reorg fixture ui files and add readme image
