+++ pkg/editor/fixturesPanel.go
@@ -139,10 +139,10 @@ func updateArray(fixtures []fixture.Fixture) [][]string {
 	return data
 }
 
-func NewFixturePanel(sequences []*common.Sequence, w fyne.Window, group int, number int, fixtures *fixture.Fixtures, commandChannels []chan common.Command) (popupFixturePanel *widget.PopUp, err error) {
+func NewFixturesPanel(sequences []*common.Sequence, w fyne.Window, group int, number int, fixtures *fixture.Fixtures, commandChannels []chan common.Command) (popupFixturePanel *widget.PopUp, err error) {
 
 	if debug {
-		fmt.Printf("NewFixturePanel\n")
+		fmt.Printf("NewFixturesPanel\n")
 	}
 
 	fp := FixturesPanel{}
